Automatic Resume Analysis/Resume Analysis ATS Score/Step-3 resume matching/requirements.txt
nltk
PyMuPDF
python-docx



# project/
# ├── Text_Preprocessing.py
# ├── requirements.txt
# ├── pdf_resumes/
# │   ├── resume1.pdf
# │   └── resume2.pdf
# ├── docx_resumes/
#     ├── resume1.docx
#     └── resume2.docx
